+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 1.1.0:
+  - added metrics to track strategy operation results
   - provide release metric in `vouch_release`
   - provide ready metric in `vouch_ready`
   - handle chain reorganisations, updating duties as appropriate